requirements.txt
colorlog==6.9.0
homeassistant==2025.2.4
pip>=21.3.1
ruff==0.12.2
python-rako-2025==0.3.10
